Golang Hello Script -mv misplaced file
